Added getting and saving bitmap preview images. Moved IPictureDisp but kept it in its own namespsace.
